Step 5 without verif
add every if and Block of name

+++ install-debug.sh
@@ -31,7 +31,9 @@ Mongodb=lien
 
 
 
-#####################################################################
+########################################################################
+#	                    UPDATE AND UPGRADE		               #
+########################################################################
 #DEFAULT SOURCES LIST OF DEBIAN BUSTER 
 echo add default buster source.list
 sudo cp sources.list /etc/apt/sources.list
@@ -175,7 +177,6 @@ sudo apt-get update
 sudo apt-get upgrade
 
 ########################################################################
-########################################################################
 #	                 	C and C++ 		               #
 ########################################################################
 #C and C++ 
@@ -189,7 +190,6 @@ then
 	sudo apt-get install g++	
 fi
 
-########################################################################
 ########################################################################
 #	      		        Python		             	       #
 ########################################################################
@@ -205,8 +205,6 @@ then
 
 fi
 
-
-########################################################################
 ########################################################################
 #	                 	   NODE	                               #
 ########################################################################
@@ -237,7 +235,10 @@ then
 	# Angular FullStack - My favorite MEAN boilerplate (MEAN = MongoDB, Express, Angularjs, Nodejs)
 	sudo npm install -g generator-angular-fullstack generatsudo 
 	sudo apt-get install gitor-angular-fullstack
+	echo NODE installed
 
+else
+	echo no NODE
 fi
 
 #Update
@@ -254,8 +255,8 @@ then
 	sudo apt-get install libssl-dev 
 	sudo apt-get install libffi-dev 
 	#with Python
-	pip install cryptography
-
+	sudo pip install cryptography
+	echo cryptography libs installed
 else
 	echo no cryptography libs
 
@@ -264,7 +265,7 @@ fi
 
 
 ########################################################################
-#	                 	Lib OPENGL 	                       #
+#	                 	Libs usefull	                       #
 ########################################################################
 # Installing crypto packagess
 read -p "OPENGL lib ? yes or no " -n 1 -r
@@ -273,15 +274,16 @@ then
 	
         sudo apt-get install libgl1
         sudo apt-get install libgl1-mesa-dev 
-
+	sudo apt-get install lib32ncurses5-dev 
+	sudo apt-get install lib32z1-dev
+	sudo apt-get install lib32stdc++6
+	echo OPENGL and lib32 installed
+	
         
 else
 	echo no OPENGL libs
 
 fi
-
-
-########################################################################
 
 ########################################################################
 #	                   Install texteditor   	               #
@@ -301,9 +303,10 @@ then
 		cp vim/.* ~/
 		curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
 	    	https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-	else
-		echo no vim config
+		echo VIM : ~/.vim >> locate.txt 
 	fi
+else
+	echo no vim config
 fi
 
 sudo apt-get update
@@ -314,12 +317,18 @@ sudo apt-get upgrade
 #	                   Install ZSH config		               #
 ########################################################################
 # Zsh
-sudo apt-get install -y zsh
-sh -c "$(curl -fsSL $ZSH)"
-#cp config
-sh zsh/fonts/install.sh 
-cp zsh/.* ~/
-
+read -p "Install zsh and config with Oh-My-Zsh ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	sudo apt-get install zsh
+	sh -c "$(curl -fsSL $ZSH)"
+	#cp config
+	sh zsh/fonts/install.sh 
+	cp zsh/.* ~/
+	echo ZSH : ~/.zsh >> locate.txt 
+else
+	echo no zsh config
+fi
 
 ########################################################################
 #	                  Install usefull tools 	               #
@@ -344,13 +353,47 @@ then
         sudo apt-get install curl 
 	sudo apt-get install git
 	sudo apt-get install filezilla
+	sudo apt-get install smartmontools 
+	sudo apt-get install ethtool
+	sudo apt-get install -y thunderbird
+	#dual boot
+	sudo apt-get install -y grub-customizer
+	#  Notes
+	sudo apt-get install -y xpad
+	sudo apt-get install -y basket
+	
+	#Usually installed by default 
+	#sudo apt-get install libreoffice
+	#sudo apt-get install wget
+	#sudo apt-get install firefox
+	#sudo apt-get install ssh
+	#sudo apt-get install file
 
-#Usually installed by default 
-#	sudo apt-get install wget
-#	sudo apt-get install firefox
-#       sudo apt-get install ssh
-#       sudo apt-get install file
+else
+	echo no usefull tools
+fi
 
+
+########################################################################
+#	                  Install less usefull tools 	               #
+########################################################################
+#Some random useful stuff
+read -p "Install usefull tools 2 ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	# Erlang - Actor Concurrency Model based Programming Language
+	sudo apt-get install erlang
+	# KVM acceleration and cpu checker
+	sudo apt-get install cpu-checker
+	sudo apt-get install qemu-kvm libvirt-bin ubuntu-vm-builder bridge-utils
+	sudo apt-get install virt-manager
+
+	# RPM and alien - sometimes used to install software packages
+	sudo apt-get install rpm
+	sudo apt-get install alien dpkg-dev debhelper
+
+	# Calibre - Ebook reader and converter
+	sudo -v && wget -nv -O- https://raw.githubusercontent.com/kovidgoyal/calibre/master/setup/linux-installer.py | sudo python -c "import sys; main=lambda:sys.stderr.write('Download failed\n'); exec(sys.stdin.read()); main()"
 
 else
 	echo no usefull tools
@@ -379,72 +422,59 @@ then
 	sudo apt-get install cabextract 
 	sudo apt-get install file-roller
 	
+	
 else
 	echo no archive extractor
 fi
 
 
 ########################################################################
-#	                 Install Archive extractors	               #
-########################################################################
-#Tools
-# TLP - saves battery when Ubuntu is installed on Laptops
-#sudo apt-get remove laptop-mode-tools
-#Uncomment Ubuntu
-#sudo add-apt-repository ppa:linrunner/tlp
-#Need bask port Debian (backport)
-sudo apt-get update
-
-sudo apt-get install -y tlp tlp-rdw smartmontools ethtool
-sudo tlp start
-sudo tlp stat
-
-# Erlang - Actor Concurrency Model based Programming Language
-sudo apt-get install -y erlang
-
-#  Notes
-sudo apt-get install -y xpad
-sudo apt-get install -y basket
-
-
-# KVM acceleration and cpu checker
-sudo apt-get install -y cpu-checker
-sudo apt-get install -y qemu-kvm libvirt-bin ubuntu-vm-builder bridge-utils
-sudo apt-get install -y virt-manager
-sudo apt-get install -y lib32z1 lib32ncurses5 lib32bz2-1.0 lib32stdc++6
-
-# RPM and alien - sometimes used to install software packages
-sudo apt-get install -y rpm
-sudo apt-get install -y alien dpkg-dev debhelper
-
-# Calibre - Ebook reader and converter
-sudo -v && wget -nv -O- https://raw.githubusercontent.com/kovidgoyal/calibre/master/setup/linux-installer.py | sudo python -c "import sys; main=lambda:sys.stderr.write('Download failed\n'); exec(sys.stdin.read()); main()"
-
-
-
-########################################################################
-#	                 Install Archive extractors	               #
+#	               	      Install TLP		               #
 ########################################################################
 
-# Dictionary Client and Server with Thesaurus
-sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu $(lsb_release -sc) universe"
-sudo apt-get install -y dict
-sudo apt-get install -y dictd
-sudo apt-get install -y dict-gcide
-sudo apt-get install -y dict-wn
-sudo apt-get install -y dict-devil
-sudo apt-get install -y dict-moby-thesaurus
+read -p " install TLP (saves battery for Laptops) ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	# TLP - saves battery when Ubuntu is installed on Laptops
+	sudo apt-get remove laptop-mode-tools
+	#Uncomment Ubuntu
+	#sudo add-apt-repository ppa:linrunner/tlp
+	#Need bask port Debian (backport)
+	sudo apt-get update
 
+	sudo apt-get install -y tlp tlp-rdw 
+	sudo tlp start
+	sudo tlp stat
+else
+	echo no TLP
+fi
 
 ########################################################################
-#	                 Install Archive extractors	               #
+#	                 Dictionary client server	               #
+########################################################################
+read -p "Install dictionary ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	# Dictionary Client and Server with Thesaurus
+	sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu $(lsb_release -sc) universe"
+	sudo apt-get install dict
+	sudo apt-get install dictd
+	sudo apt-get install dict-gcide
+	sudo apt-get install dict-wn
+	sudo apt-get install dict-devil
+	sudo apt-get install dict-moby-thesaurus
+else
+	echo no dictionary
+fi
+
+########################################################################
+#	                 	   Golang          	               #
 ########################################################################
 read -p "Install Golang ? yes or no" -n 1 -r
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
-sudo apt-get install golang
+	sudo apt-get install golang
 fi
-
 
 
 ########################################################################
@@ -518,91 +548,152 @@ fi
 
 
 
-# Docker
-sudo apt-get install docker -y
-curl -fsSL get.docker.com -o get-docker.sh
-sh get-docker.sh
-# Docker-compose
-sudo pip install docker-compose -y
+########################################################################
+#	           	          DOCKER			       #
+########################################################################
+read -p "Install Docker ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	# Docker
+	sudo apt-get install docker
+	cd ~/temp
+	curl -fsSL get.docker.com -o get-docker.sh
+	sh get-docker.sh
+	cd -
+	# Docker-compose
+	sudo pip install docker-compose -y
+	
+	sudo apt-get update && sudo apt-get install -y apt-transport-https curl
+	sudo curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
+	cat <<EOF >/etc/apt/sources.list.d/kubernetes.list
+	deb https://apt.kubernetes.io/ kubernetes-xenial main
+	EOF
+	sudo apt-get update
+	sudo apt-get install -y kubelet kubeadm kubectl
+	sudo apt-mark hold kubelet kubeadm kubectl
+fi
+
+########################################################################
+#	           	          Multimedia			       #
+########################################################################
+read -p "Install multimedia ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#Multimedia
+	sudo apt-get install -y vlc
+	sudo apt-get install -y kazam
+	sudo apt-get install -y clementine
+fi
+
+########################################################################
+#	           	         Onenote		 	       #
+########################################################################
+read -p "Install Onenote ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#Onenote
+	#With npm
+	sudo npm install -g p3x-onenote --unsafe-perm=true --allow-root
+	p3x-onenote &
+fi
+
+########################################################################
+#	           	         Teamviewer		 	       #
+########################################################################
+read -p "Install Teamviewer ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#Teamviewer
+	cd ~/temp && wget -O teamviewer.deb https://download.teamviewer.com/download/linux/teamviewer_amd64.deb
+	sudo dpkg -i teamviewer.deb
+	sudo apt-get install -f
+fi
+
+########################################################################
+#	           	         Dropbox		 	       #
+########################################################################
+read -p "Install Dropbox ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#dropbox
+	#URL 09/2019
+	#cd ~/temp && wget -O dropbox.deb https://www.dropbox.com/download?dl=packages/ubuntu/dropbox_2019.02.14_amd64.deb
+	#sudo ~/temp/.dropbox-dist/dropboxd
+	cd ~/temp
+	git clone git://github.com/dropbox/dropbox-sdk-python.git
+	cd dropbox-sdk-python
+	sudo python setup.py install
+fi
 
 
-#Tool ++
-sudo apt-get install -y vlc
-sudo apt-get install -y thunderbird
-sudo apt-get install -y libreoffice
-sudo apt-get install -y locate
-sudo apt-get install -y kazam
-sudo apt-get install -y clementine
-#dual boot
-sudo apt-get install -y grub-customizer
+
+########################################################################
+#	           	       Sublimtext			       #
+########################################################################
+read -p "Install Sublimetext ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#sublimText
+	#URL 09/2019
+	cd ~/temp && wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo apt-key add -
+	sudo apt-get update
+	sudo apt-get install sublime-text
+fi
 
 
-#Onenote
-#With npm
-sudo npm install -g p3x-onenote --unsafe-perm=true --allow-root
-p3x-onenote &
+########################################################################
+#	           	          VsCode			       #
+########################################################################
+read -p "Install VSCODE ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#VsCode
+	#URL 09/2019
+	cd ~/temp && wget -O vscode.deb https://go.microsoft.com/fwlink/\?LinkID\=760868 && sudo dpkg -i vscode.deb
+	sudo apt-get install -f
+fi
 
+########################################################################
+#	           	          LAMPP				       #
+########################################################################
+read -p "Install multimedia ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	#LAMPP:
+	#Mysql
+	wget http://repo.mysql.com/mysql-apt-config_0.8.13-1_all.deb
+	sudo dpkg -i mysql-apt-config_0.8.13-1_all.deb
+	sudo apt-get install -f
 
-#Teamviewer
- cd ~/temp && wget -O teamviewer.deb https://download.teamviewer.com/download/linux/teamviewer_amd64.deb
-sudo dpkg -i teamviewer.deb
-sudo apt-get install -f
+	#Workbench 
+	wget -O workbench.deb https://dev.mysql.com/get/Downloads/MySQLGUITools/mysql-workbench-community_8.0.17-1ubuntu18.04_amd64.deb
 
+	#Apache server
+	#sudo apt-get install apache2 apache2-doc apache2-npm-prefork apache2-utils libexpat1 ssl-cert -y
 
-#dropbox
-#URL 09/2019
-#cd ~/temp && wget -O dropbox.deb https://www.dropbox.com/download?dl=packages/ubuntu/dropbox_2019.02.14_amd64.deb
-#sudo ~/temp/.dropbox-dist/dropboxd
-cd ~/temp
-git clone git://github.com/dropbox/dropbox-sdk-python.git
-cd dropbox-sdk-python
-sudo python setup.py install
+	#Install PHP (php7.0 latest version of PHP)
+	#sudo apt-get install libapache2-mod-php7.0 php7.0 php7.0-common php7.0-curl php7.0-dev php7.0-gd php-pear php-imagick php7.0-mcrypt php7.0-mysql php7.0-ps php7.0-xsl -y
 
-#sublim
-#URL 09/2019
-cd ~/temp && wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo apt-key add -
-sudo apt-get update
-sudo apt-get install sublime-text
+	#Install Phpmyadmin(for database)
+	#sudo apt-get install phpmyadmin
+fi
 
+########################################################################
+#	           	       MongoDB				       #
+########################################################################
+read -p "Install multimedia ? yes or no" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	# Mongodb, Installing and starting server
+	sudo apt install gnupg
+	cd ~/temp
+	wget -qO - https://www.mongodb.org/static/pgp/server-4.2.asc | sudo apt-key add -
+	echo "deb http://repo.mongodb.org/apt/debian stretch/mongodb-org/4.2 main" | sudo tee /etc/apt/sources.list.d/mongodb-org-4.2.list
+	sudo apt-get update
+	sudo apt-get install -y mongodb-org
 
-
-
-#VsCode
-#URL 09/2019
-cd ~/temp && wget -O vscode.deb https://go.microsoft.com/fwlink/\?LinkID\=760868 && sudo dpkg -i vscode.deb
-sudo apt-get install -f
-
-
-
-#LAMPP:
-#Mysql
-wget http://repo.mysql.com/mysql-apt-config_0.8.13-1_all.deb
-sudo dpkg -i mysql-apt-config_0.8.13-1_all.deb
-sudo apt-get install -f
-
-#Workbench 
-wget -O workbench.deb https://dev.mysql.com/get/Downloads/MySQLGUITools/mysql-workbench-community_8.0.17-1ubuntu18.04_amd64.deb
-
-#Apache server
-#sudo apt-get install apache2 apache2-doc apache2-npm-prefork apache2-utils libexpat1 ssl-cert -y
-
-#Install PHP (php7.0 latest version of PHP)
-#sudo apt-get install libapache2-mod-php7.0 php7.0 php7.0-common php7.0-curl php7.0-dev php7.0-gd php-pear php-imagick php7.0-mcrypt php7.0-mysql php7.0-ps php7.0-xsl -y
-
-#Install Phpmyadmin(for database)
-#sudo apt-get install phpmyadmin
-
-
-
-# Mongodb, Installing and starting server
-sudo apt install gnupg
-cd ~/temp
-wget -qO - https://www.mongodb.org/static/pgp/server-4.2.asc | sudo apt-key add -
-echo "deb http://repo.mongodb.org/apt/debian stretch/mongodb-org/4.2 main" | sudo tee /etc/apt/sources.list.d/mongodb-org-4.2.list
-sudo apt-get update
-sudo apt-get install -y mongodb-org
-
-
+fi
+#########################################################################
 
 sudo apt-get update -y
 sudo apt-get upgrade -y
